backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.1
pydantic==2.4.2
pydantic-settings==2.0.3
groq==0.4.1